requirements.txt
django
mysqlclient
django_mysql
djangorestframework
django-cors-headers
django-webpack-loader
django-rest-knox